+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ CHANGELOG
 ===========
 
  * fixed sqldump to display stdout on stderr allowing it to be piped correctly
+ * added gondor env and env:set to view and set instance environment variables
 
 1.0.1
 =====